// frontends/epro/src/app/api/model/cell-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FieldModelDTO } from './field-model-dto';
import { VisibilityCriteriaDTO } from './visibility-criteria-dto';


/**
 * A cell in the CRF display. Can contain a field or static information. Can be displayed conditionally based on specified criteria.
 */
export interface CellDTO { 
    id: string;
    datasetModelId: string;
    fieldModelId: string;
    textBefore?: { [key: string]: string; };
    textAfter?: { [key: string]: string; };
    visibilityCriteria: Array<VisibilityCriteriaDTO>;
    displayLabel: boolean;
    displayPossibleValueLabels: boolean;
    possibleValuesColumnNumber?: number;
    possibleValuesColumnWidth?: number;
    colspan?: number;
    hasPrintButton: boolean;
    printButtonLabel?: { [key: string]: string; };
    addEventButton?: string;
    addEventButtonLabel?: { [key: string]: string; };
    fieldModel?: FieldModelDTO;
    cssCodeForLabel?: string;
    cssCodeForInput?: string;
    cssCode?: string;
}