// frontends/epro/src/app/api/model/scope-model-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ScopeModelDTO { 
    id: string;
    shortname: { [key: string]: string; };
    longname?: { [key: string]: string; };
    description?: { [key: string]: string; };
    pluralShortname: { [key: string]: string; };
    parentIds: Array<string>;
    defaultParentId: string;
    childScopeModelIds: Array<string>;
    root: boolean;
    leaf: boolean;
    virtual: boolean;
    datasetModelIds: Array<string>;
    formModelIds: Array<string>;
    workflowIds: Array<string>;
    defaultProfileId: string;
}