// frontends/epro/src/app/api/model/study-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DatasetModelDTO } from './dataset-model-dto';
import { EventModelDTO } from './event-model-dto';
import { FormModelDTO } from './form-model-dto';
import { WorkflowDTO } from './workflow-dto';
import { LanguageDTO } from './language-dto';
import { CountryDTO } from './country-dto';
import { ScopeModelDTO } from './scope-model-dto';
import { MenuDTO } from './menu-dto';
import { ProfileDTO } from './profile-dto';


export interface StudyDTO { 
    id: string;
    shortname: { [key: string]: string; };
    color: string;
    url: string;
    homePage: MenuDTO;
    eproEnabled: boolean;
    introductionText?: string;
    copyright?: string;
    activatedLanguages: Array<LanguageDTO>;
    defaultLanguage: LanguageDTO;
    environment: StudyDTO.EnvironmentEnum;
    leafScopeModel: ScopeModelDTO;
    email: string;
    clientEmail: string;
    welcomeText: string;
    configUser: string;
    configDate: Date;
    scopeModels: Array<ScopeModelDTO>;
    eventModels: Array<EventModelDTO>;
    datasetModels: Array<DatasetModelDTO>;
    formModels: Array<FormModelDTO>;
    workflows: Array<WorkflowDTO>;
    profiles: Array<ProfileDTO>;
    menus: Array<MenuDTO>;
    countries: Array<CountryDTO>;
    userAttributes: Array<string>;
    rootScopeModelId: string;
    leafScopeId: string;
    leafScopeModelIds: Array<string>;
}
export namespace StudyDTO {
    export type EnvironmentEnum = 'PROD' | 'VAL' | 'DEV';
    export const EnvironmentEnum = {
        PROD: 'PROD' as EnvironmentEnum,
        VAL: 'VAL' as EnvironmentEnum,
        DEV: 'DEV' as EnvironmentEnum
    };
}